rename mockLocation.ts to buildLocation.ts
