+++ packages/core/test/emulate/buildLocation.ts
