+++ packages/core/test/index.ts
@@ -5,7 +5,7 @@ export * from './cookie'
 export * from './registerCleanupTask'
 export * from './interceptRequests'
 export * from './emulate/createNewEvent'
-export * from './emulate/mockLocation'
+export * from './emulate/buildLocation'
 export * from './emulate/mockClock'
 export * from './emulate/mockReportingObserver'
 export * from './emulate/mockZoneJs'
